test: consolidate per-project sequence tests in config/sequence.test.ts
Move the per-project sequence serialization test out of configureVitest.test.ts
and the inheritance test out of sequence-shuffle.test.ts into a single
runInlineTests case under config/sequence.test.ts. One root config + three
projects covers project override, non-inheritance for bare inline projects,
inheritance via `extends: true`, and the shared root seed.
