+++ test/e2e/test/config/sequence-shuffle.test.ts
@@ -162,51 +162,3 @@ test('should not log seed when shuffle is disabled', async () => {
 
   expect(stdout).not.toContain('Running tests with seed')
 })
-
-test('per-project sequence is inherited from the root config only with `extends`', async () => {
-  const { ctx, stderr } = await runInlineTests({
-    'vitest.config.ts': /* ts */ `
-      import { defineConfig } from 'vitest/config'
-      export default defineConfig({
-        test: {
-          sequence: { concurrent: false, hooks: 'list', setupFiles: 'list', seed: 123 },
-          projects: [
-            { test: { name: 'no-extends', include: ['a.test.ts'] } },
-            { extends: true, test: { name: 'with-extends', include: ['a.test.ts'] } },
-          ],
-        },
-      })
-    `,
-    'a.test.ts': /* ts */ `
-      import { test } from 'vitest'
-      test('example', () => {})
-    `,
-  })
-
-  expect(stderr).toBe('')
-  const pick = (name: string) => {
-    const { concurrent, hooks, setupFiles, seed } = ctx!.projects.find(p => p.name === name)!.serializedConfig.sequence
-    return { concurrent, hooks, setupFiles, seed }
-  }
-  expect({
-    // a bare inline project does not inherit the root `sequence`; it uses defaults
-    'no-extends': pick('no-extends'),
-    // `extends: true` inherits the root `sequence`
-    'with-extends': pick('with-extends'),
-  }).toMatchInlineSnapshot(`
-    {
-      "no-extends": {
-        "concurrent": undefined,
-        "hooks": "stack",
-        "seed": 123,
-        "setupFiles": undefined,
-      },
-      "with-extends": {
-        "concurrent": false,
-        "hooks": "list",
-        "seed": 123,
-        "setupFiles": "list",
-      },
-    }
-  `)
-})
